Switch to the `/authenticate` endpoint
See https://singly.com/docs/authorization

`/authenticate` is replacing `/authorize` and discarding the use of cookies on
api.singly.com to merge accounts. The developer must now attach an
`access_token` parameter to the auth URL if the intent is to connect a new
profile to an account. Without an `access_token`, a new account will be created
(or the account will just be logged into if it exists).